Added new scripts

+++ scripts/gen_env.py
@@ -10,18 +10,19 @@ true_norm_name = params['true_norm']['name']
 true_norm_exp = params['true_norm']['exp']
 colour_specific = params['colour_specific']
 shape_specific = params['shape_specific']
-target_area = params['target_area']
+target_area_str = params['target_area']
 
-print(f'True norm name: {true_norm_name}')
-print(f'True norm exp: {true_norm_exp}')
-print(f'Target area: {target_area}; {type(target_area)}')
+print(type(colour_specific))
 
+#
+#
+#
 
-#colour_specific: ['r', 'g', 'b']
-#shape_specific: ['square', 'circle', 'triangle']
-#target_area: -0.8, 0.7 ; 0.25, 0.99
-
-if true_norm_name = 'default':
+if (true_norm_name == 'default' and 
+    colour_specific == ['r', 'g', 'b'] and
+    shape_specific == ['square', 'circle', 'triangle'] and
+    target_area_str.replace(' ', '') == '-0.8, 0.7 ; 0.25, 0.99'.replace(' ','')
+   ):
     shutil.copyfile('data/default_env.pickle', 'data/env.pickle')
 else:
     print('To do: generate new env')
